Here is the GUI
Git is really, REALLY annoying. Here is a mostly working GUI. I tried to
code all the methods so it'll be easy to modify when the server code is
added. The names and all associated text still needs work. Per usual,
the eclipse xml stuff puts the text boxes in weird places. All card
stuff is definitely working.