Good working copy of the GUI
I fixed a lot of small issues with the XML file. Now the name fields
span the entire screen. The Gameplay file also has small changes. The
big and small blinds work well. The bet button also works and stores it
in a global variable.
Skye this should work well as a GUI for you to put your code into.