Got the notification parser working with ||
I modified the GCMintentservice to recognize when a notification with a
code in it comes down. You must send your message with the format:
||1||Geoff||
The codes are as follows:
When player logs in send message to all other players and display the
new player on their screen.
Code 1, Name

send message when the game starts and have android send a get request
for its 2 cards
Code 2, Null

send a message when it is your turn to bet.
Code 3, Null

send a message when the flop goes and have the phone send a get request
for the flop
Code 4, Null
same with the turn and river cards.
Code 5, Null
Code 6, Name of Final card

send a message with winner.
Code 7, Name of winner

There is also error checking in this. If something comes down wrong, the
notification still displays, only as "Switch case isn't working" as the
notification text.